refactoring

+++ lc3common.h
@@ -6,7 +6,7 @@
 
 int is_register(char *token);
 int is_imm5(char *token, int *imm5);
-int is_PCoffset11(char *token, int *label);
+int is_PCoffset11(char *token, long *label);
 char **instruction_tokens(char *asm_instr, char *instr_name, int num_tokens);
